requirements.txt
google-generativeai
beautifulsoup4
requests
python-dotenv
trafilatura
